Redefine most data types involved in scoring and sequencing.
Types for translating scored notes to sequenced notes come in two flavors: tonal notes and drums. The major difference between the two is that that a drum kit is modeled as a single instrument, whereas every tonal instrument has it's own sound kit -- said another way, the drum kit is like a default instrument where notes map to sounds rather than frequencies, whereas tonal instruments are drum kits with a label and a mapping from keyboard keys to frequencies. Still, there is a lot of overlap between the two flavors. This commit includes changes that build scores and note sequences out of the same data types but parameterized over how the sounds for these instruments are indexed, either by a keyboard key or by a drum key.