Create test suites, move test functions from TinyRelDB into suite.
I created the test suite in the Cabal file so now the `cabal test`
command does the right thing, and the tests are no longer built in to
the ProcGen library, which is good because they need not be there
taking up space. I had to export VarWord35 and associated functions
from the ProcGen.TinyRelDB module in order for the tests to be moved
into another module where these data types are imported.