Create the UTF8Char primitive to be used in place of 'Prelude.Char'.
Previously, 'Prelude.Char' was instantiating the typeclasses in the
"ProcGen.TinyRelDB" module. But the serialized format of the data type
which I defined in this module may have a different serialized format
than what the ordinary 'Prelude.Char' instance for the
'Data.Binary.Binary' typeclass has. So it is best to create a newtype
which instantiates these typeclasses with the format that I defined
myself.

I also created the rest of the primitive type instances for the
'PackedVectorElem' typeclass.