Implemented homogenous vector primitives, delete 'PackedVectorElem'.
It is finally possible to store a homogenous vector as a primitive in
a tagged database row. It is necessary to first encode the vector as a
'PackedVector' data type, which is a primitive type for any element
that instntiates 'IsNumericPrimitive'.

Although I would like to have a separate typeclass for 'PackedVector'
elements so that arbitrary packed elements can be stored into a
'PackedArray', it isn't possible to encode tags for arbitrary data
types in a 'TaggedRow', so I had to restrict the class of data types
that can be encoded to a 'PackedVector' to the class of types that
instantiate 'IsNumericPrimitive'.

Also, ByteStrings are now primitives tagged as blobs and not strings.
Before this commit both lazy and strict byte strings were tagged with
the 'UTF8String' primitive type tag. Now they are tagged with the
'BinaryBlob' type tag. To serialize a UTF8-encoded byte string with a
'UTF8String' tag, there are two new APIs to do this:
'putPrimLazyUTF8String' and the strict counterpart to this put
function, and also the 'getPrimLazyUTF8String' along with the strict
counterprat to this get function.