Bug fix: 'rowElem' function must account for TaggedRow header size.
Also, I realized the 'PrematureEndOfRow' data constructor must never
take an "expected" tag, because this exception is only thrown when
there are no values expected! If anything the 'corruptExpectedTag'
belongs in the 'CorruptRow' data constructor.

Also redefined 'readRowPrim' to reuse the 'unpackRowElem', and made
sure the 'currentTaggedRow' function properly checks for
'Data.Binary.Get' decoder failures and evaluates to 'mzero' on
failure. It may be better to let the exception propagate here, I am
not sure.

Also, I 'isolate'd the UTFChar decoder in the 'getUTFChar' function in
order to make the decoder atomic, so it will throw the correct
exception if there are not exactly 3 bytes in the parsed ByteString.