Instantiate SelectEnv into Show, create constructor for 'SelectEnv'.
I got rid of the 'showHeaderIfCorruptRow' function because it seems to
be better to construct your own 'SelectEnv' and show this
'SelectEnv' (including the header and the hex dump of it's contents)
in the case that an exception is thrown. So the code for
'showHeaderIfCorruptRow' is now the code used to instantiate
'SelectEnv' into the 'Show' typeclass.