Bug fix: 'calcByteSize' must NOT use 'Data.Binary.put'.
The 'Data.Binary.put' function has it's own format for primitive
types. For example, 'Float' and 'Double' values both require 13 bytes
of storage space, rather than the standard 4 and 8
bytes (respectively) of storage space. It was necessary to change the
definition of 'calcByteSize' to take a continuation function of type
'Data.Binary.Put.Put' which uses the correct format for the given
primitive type, for example 'Data.Binary.Put.putFloathost' for
computing the true size of a 'Float' value.