Make (Int64 -> Int) overflow checking consistent.
I am doing this in the hypothetical situation that this library may be
compiled on 32-bit systems, and conversion between Int64 and Int may
result in an overflow. But the fact that array and string element
sizes are stored as VarWord35 values, this may be a moot point
entirely, I may need to rewrite everything 64-bit clean, even when
compiling on 32-bit systems, which means using lazy ByteStrings for
absolutely everything, even in situations where it seems a strict
ByteString is more appropriate. I'll have to sit and think about this
a bit more later, for now overflows are checked and exceptions with
somewhat helpful error messages are thrown.