Bug fix, ProcGen.Music.WaveFile: explicitly open/close file handles.
The 'readFile' and 'writeFile' functions in the "Data.ByteString.Lazy"
module don't always close their file handles after fully evaluating,
even when using the 'Control.Exception.evaluation' function, at least
this was the behavior when testing in GHCi. This was leading to
exceptions being thrown when the same file is read and written many
times in rapid succession. I rewrote these functions to make use of
'hReadWave' and 'hWriteWave' instead so that I could manage the
opening and closing of handles myself.