Define FDSignal base in terms NormFrequency and a FreqCoeficient.
In order to make it easier to compute a 'FDSignal' in terms of it's
base frequency, changed the base frequency component of the 'FDSignal'
to be defined in terms of a normalized frequency value NormFrequency,
which stores a value between 0 and 1 and scales up to an actual
audible frequency when extracted. Similarly, component frequencies are
now defined in terms of coeficients of this base frequency. In order
to make sure I got the math right from these changes, I introduced 2
new types to ensure these values are properly computed where they are
used, much like assigning dimensional units to values in your
computations. I'd like to do this for all numerical types at some
point.